+++ Changelog.md
@@ -1,3 +1,10 @@
+# Version 1.1.0
+
+- Ui Resdesign for the options frame
+- Logo added to header of the options frame
+- Version Checker
+- Buttons added for opacity options, this is coming in a future update. 
+
 # Version 1.0.12
 
 - TOC Bump
@@ -5,6 +12,7 @@
 # Version 1.0.11
 
 - Fixed issue with version on options frame
+
 
 # Version 1.0.10